Creating JUDOConfig class which manages loading for default and user-defined judo.properties files.
In addition, a default JUDO properties file is archived with the JAR. The user properties file is
then "merged" with the default properties file, giving precedence to the user defined configurations.

This removes complexity from JUDOIDE and fixes bugs with how the previous implementation handles
default configuration values.

Finally, this sets the default JUDO programs directory to ~/MyJUDOPrograms, instead of
<JUDO install location>/MyJUDOPrograms . Further work is needed to open and save .judo
files from this directory.