Incomplete refactor to abstract details away from JUDOIDE.java .
This introduces the notion of a JUDOProgram (a nearly atomic
representation of a JUDO program). This contains methods for setting and
retrieving metadata (about the JUDO version used to create it, app type,
and app name). This also includes code to auto-update the metadata to
match the users' code whenever it is updated.

The JUDOProgramManager is a new class that abstracts file system
operations away from JUDOIDE. It handles common operations such as save,
open, delete, and create. It is incomplete and has many stubs. The only
completed function is the openProgram method - more work here is
required.

The JUDOCompiler is an early step at abstracting away unnecesary
compiler details from JUDOIDE. The CompilerError and Command class
could both be integrated and perhaps replaced by this class.

The template code "JUDOApp.template" was moved from the org/judo
directory to be a static resource packaged with the JAR.
See JUDOCompiler for more information.

The FilenameFilter used to identify JUDO files (.judo) was moved from
JUDOIDE to a new static nested class in JUDOUtils. A single static
instance is lazily created and stored, which can be retrieved using
JUDOUtils.getJUDOProgramFilenameFilter .